[#32] Create function to convert model to json
