+++ lib/ui/questions/survey_question_view.dart
@@ -29,7 +29,7 @@ class _SurveyQuestionViewState extends ConsumerState<SurveyQuestionView> {
       nextPage();
     });
     return PageView.builder(
-      // physics: const NeverScrollableScrollPhysics(),
+      physics: const NeverScrollableScrollPhysics(),
       controller: _pageController,
       onPageChanged: (index) {
         widget.onPageChanged.call(index);
